FIX:verify exiftool_write before trying to delete related tmpfile [notag]
git-svn-id: http://svn.resourcespace.com/svn/rs/trunk@11814 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ pages/download.php
@@ -312,12 +312,9 @@ if(!hook('replacefileoutput'))
 
 // Deleting Exiftool temp File:
 // Note: Only for downloads (not previews)
-if('' == $noattach && -1 == $alternative)
+if('' == $noattach && -1 == $alternative && $exiftool_write && file_exists($tmpfile))
     {
-    if(file_exists($tmpfile))
-        {
-        delete_exif_tmpfile($tmpfile);
-        }
+    delete_exif_tmpfile($tmpfile);
     }
 
 hook('beforedownloadresourceexit');
